Fixed some missing API functions.
Some definitions in the spec did not have a `x-kubernetes-group-kind-version`
but there were operations that had a `x-kubernetes-group-kind-version` for them.
These operations were not getting emitted as API functions. Fixed by manually
setting the missing `x-kubernetes-group-kind-version` on the definitions.

Also added an assert that all API functions are emitted.

Also order API functions in order of group-version-kind instead of
group-kind-version.