Implemented conditional compilation macros that emit differently based on enabled features.
- The `k8s_if_{version}` macros emit their contents if the corresponding feature
  is enabled.

- The `k8s_if_le_{version}` macros emit their contents if the corresponding
  or a lower feature is enabled.

- The `k8s_if_ge_{version}` macros emit their contents if the corresponding
  or a higher feature is enabled.

- The `k8s_match` macro emits a match expression. The arms of the match
  expression can be wrapped in the above conditional compilation macros.

This makes it so that library crates (ie not end-user crates) don't have to
declare their own features to forward to k8s-openapi crates. Instead they can
use these macros and leave the feature selection to the final binary crate that
needs to talk to a cluster, which it can do by having its own dependency on
the k8s-openapi crate to just enable specific features.

This is required because cargo does not give a way to have `#[cfg()]` attributes
based on features of *other* crates in the dep tree, and redeclaring features
that forward to a dependency's features is both tedious and error-prone.
The latter is because two crates can enable two different features of the same
underlying crate without knowing of each other, and would thus become
inconsistent.

Fixes #16