Removed support for v1.7
The different imports for v1.7 were a common reason to use `k8s_*` macros.
Now that only a single feature is supported, a lot of the uses of the macros
went away, except for this one.

None of the major clouds appear to support it, so that's not a reason to
keep it either.