Elid an unnecessary bounds check when parsing string responses.
The `std::str::from_utf8` function unfortunately does not return a `&str` for
the bytes that *did* pass validation. So all the `type: string` responses have
to use `std::str::from_utf8` again with a subslice of the original slice up to
the last valid index. To avoid having to re-validate this slice, they use
the unsafe `std::str::from_utf8_unchecked` function instead.

Since it's already unsafe, there's also an opportunity to elid the
bounds checking around subslicing the original slice to the index of the first
invalid byte.