Merge watch and watchlist operations into a single watch operation.
It did not make much sense to have a watch operation with a required
`field_selector` parameter, since there was no way to guarantee that the user
set it in a way that only matched one item. From the perspective of parsing the
response, it also did not make sense to differentiate between the two kinds of
operations, since they both return `WatchEvent`s.

With this change, the original watch operations are no longer emitted.
The watchlist operations are now emitted with the same name as the original
watch operations, and with an optional `field_selector` parameter.

Eg, `Pod::watch_namespaced_pod` now has an optional `field_selector` parameter
and `Pod::watch_namespaced_pod_list` no longer exists. Whether it returns events
for a single pod or a list of pods depends on how the user set `field_selector`.
It returns a stream of `WatchEvent` in either case.

This change also reinstates the optional `field_selector` parameter for list
operations. These were incorrectly removed as part of synthesizing watch and
watchlist operations from list operations. List operations can still use
this parameter to filter the list of items.