Add `Content-Type` header to all requests with bodies.
All operations with requests bodies, except for patch requests, now add
a `Content-Type: application/json` header to the request.

The `Patch` type is now generated as an enum of `Json`, `Merge` and
`StrategicMerge` variants, each wrapping a `serde_json::Value`.

Patch operations use the variant of the `Patch` value to set the appropriate
`Content-Type` header.

This commit also fixes the response types of delete-collection operations to
contain the list type of the associated type, not the associated type itself.

Ref #43
Fixes #49