Emit subresources field in #[derive(CustomResourceDefinition)] structs.
This commit adds a new `has_subresources` meta-item of
the `#[custom_resource_definition]` attribute. If set, the generated
custom resource will have a `subresources` field of
the `apiextensions::CustomResourceSubresources` type. The precise namespace
that the type is referenced from is determined by the value of the meta-item.

Despite being a field of the CR struct, the inner fields of the value are
flattened into the CR value when (de)serializing, as they should be.

Also change the k8s-openapi-tests client to perform one read of
the response body before trying to parse it. This ensures that bodies
corresponding to the `Other(Option<serde_json::Value>)` variant are fully
parsed and printed in case of errors.

Fixes #62