Make resource metadata non-optional.
The `metadata` field is generally required, so having it be optional
required awkward code to handle it not being set even when it ought to be set.

For situations like PATCH requests, the field is indeed not required,
but setting it to an empty object `{}` via `Default::default()` also works
the same way, so there is no loss of functionality.