Use default-features=false for all deps.
This also disables the extra-traits feature of syn, since it was being used for
one `#[derive(Debug)]` that can be done without, and
one `!=` involving `AttrStyle` that can be done with `if let` instead.

base64, chrono and serde_json have "alloc" and "std" features where one of them
is required. In all three cases, it's possible to have both features enabled
and that's essentially treated the same as if only the std feature is enabled.
So this commit enables the "alloc" feature, and if anything else in
the end user's dep tree enables the "std" feature, that's fine.

The one noticeable breaking change is that enabling the schemars feature of
k8s-openapi no longer enables the derive feature of schemars, so users who want
to use `#[derive(schemars::JsonSchema)]` in their own code will need to add
an explicit dependency on schemars to enable that feature.

Ref: https://github.com/kube-rs/kube-rs/issues/650#issuecomment-946029033