Allow deserializing JSON null as an empty ByteString.
The API server allows binary data values in ConfigMaps and Secrets to be
null if validation is disabled. This commit allows such a value to be
deserialized as an empty ByteString. This was done rather than change
Secret::binary_data and ConfigMap::binary_data to be maps of Option<ByteString>
since that would be worse for ergonomics and because these values can only be
created with validation disabled.

Mapping to an empty ByteString has precedence with the API server's handling
of nulls in ConfigMap::data, which again requires validation to be disabled,
and in which case the API server itself maps the nulls to empty strings.

Fixes #114