Reorganize lib.rs into mods because it has become very long.
To avoid clobbering names from the Kubernetes spec that would be in scope
from the version module glob re-exports, the modules are prefixed with `_`.
This is only a problem for `api` right now, and the existing `deep_merge` mod
was fine already, but this commit renames all the modules just in case
it becomes a problem in the future.