Add overrides to make `api::events::v1{beta1,}::Event::event_time` optional.
This API is supposed to be a replacement for `api::core::v1::Event`, so for
backward-compatibility the `eventTime` field is returned as `null` for events
that were created using the core/v1 interface. This includes events created by
the cluster itself during bootstrap. However the spec marks this field as
required, which caused the k8s-openapi type to fail to be deserialized from
such a JSON blob.

The golang API server works around this by implementing `UnmarshalJSON` on
`MicroTime` (and `Time`) to support being deserialized from `null` -
the resulting value is the "zero" value of golang's `time.Time` struct
which represents 0001-01-01T00:00:00.000000Z

In #137 OP asked for `apimachinery::pkg::apis::meta::v1::MicroTime` to
also support being deserialized from `null` in the same way. However, testing
this revealed that there are other fields in `Event` like `reportingController`
that are also required for new events but are nevertheless marked as optional
in the OpenAPI spec.

So it seems cleaner to treat this as a spec bug that the `eventTime` field
ought to have been marked as optional, just like `reportingController` etc are,
rather than deserialize `null` to a "zero time".

Note that `reportingController` being required might possibly be a bug,
ref: https://github.com/kubernetes/kubernetes/issues/89062 , but since that
was abandoned three years ago and the "bug" hasn't been resolved, it seems
prudent to treat it as by-design for now.

Fixes #137