Remove broken test from k8s-openapi-codegen-common
3bfe7c098e6f6686d263e750605d14ef0ce6892e added the ability to handle names
containing `.` for the sake of OpenShift, but then this was inadvertently
broken by c840bb410cc32b0620b605ed58c392933372ceea . This went unnoticed
because the test added by the former was not being run.

It's not clear whether `get_rust_ident` actually needs to handle names
containing `.` or not, even for the sake of OpenShift. So this commit removes
the test for now. It also adds `cargo test` to the global test suite so that
tests added to codegen-common are not missed in the future.

Closes #144