Remove API operations-related codegen.
Updating the codegen for v1.28 is non-trivial, and nobody seems to use it
over kube anyway. It doesn't seem worth delaying v1.28 support for or
keeping around in general.

Operations are still parsed minimally because they're used to identify
resource scope.

The tests now use clientset code based on the original operations code,
which is generic over the resource type like kube but still sans-io as before.
So that is an option for users who still need the sans-io-style operations,
though it only handles common CRUD by default and not bespoke operations
like reading pod logs. If there is demand it could be released as
a separate crate.

Ref #145