Redirect docs of releases before 2022 to docs.rs
With the v0.20 release, the size of the checked out files exceeded 1GB and
GitHub Pages complained that deploying it might fail.

>Uploaded artifact size of 1091553280 bytes exceeds the allowed size of 1 GB.
>Deployment might fail.

It did actually succeed, but let's not take the risk for future releases.
So this change removes the docs for releases before 2022, ie 0.13 and earlier,
and redirects them to docs.rs