Update http to v1
reqwest still pulls in http 0.2, so change all uses of
`http::{Method, StatusCode}` to the reqwest re-exports. This is arguably
more correct since these values are eventually used with a reqwest client.

`http::{Request, Uri}` are still used because reqwest does not re-export them,
which necessitates a conversion from `http::Method` to `reqwest::Method`.
It has been marked with a `TODO(rustup)` so that it can be removed when
reqwest is updated to a version that depends on http v1.