Backport fixes from v1.30.1 to make various fields required.
v1.30.1 makes `HostAlias::ip` a required field because a different change
in v1.30.0 happened to expose this bug. However the fact that the field
is required is also true of earlier releases, so this commit adds a fixup
to all releases to do that.

Furthermore, upstream wants a bunch of `name` fields to be treated as
required, but is unable to do it for backward-compatibility so it is
allowed to be omitted and defaults to `""`. Also, just like `HostAlias::ip`,
this was only exposed in v1.30.0, but affects all releases.
Backward-compatibility is not a concern for k8s-openapi since
we already deserialize missing required String fields as `""`,
so this commit also adds fixups to make those `name` fields be required.