Add support for `no_std` in the generated code.
Crates that have `std` available also have `core` available, so types that are
available from `core` are now always referenced from `core`, such as
`core::marker::PhantomData`.

Other types like `Box` that are only available in `alloc` need to be
referenced from `alloc`, however `alloc` is not automatically available to
`std`-using programs unless they write `extern crate alloc;`. It would be
possible for k8s-openapi to have `extern crate alloc;` and then change the
codegen to refer to `alloc::boxed::Box` etc, but this would not work for
other crates that use the code generator or the derive macro. So instead
`alloc` is renamed to `std` and then the types are referenced using
`std`-relative paths like `std::boxed::Box`, which happen to work for both
`alloc` and `std`.

An alternative to this would be to have k8s-openapi re-export `alloc` and then
the generated code would reference `{local}alloc::boxed::Box` etc. This is
overkill for now so it hasn't been done, but it's an option in case
the trick of paths being the same for `alloc` and `std` doesn't work for
some types in the future.

One more thing affected by this change is having to use `.into()` to convert
from `&str` to `String` instead of `.to_owned()`, since `ToOwned` is in `alloc`
and thus not in `core`'s prelude.

This feature is only for the generated code; the code generator itself
still requires `std`.

The macro-deps test has been co-opted to test `no_std` mode.

Closes #162